// src/app/Interfaces/cadastro.ts
export interface UsuarioProp  {
    id?: number,
    nome?: string
    fone?: string
    cpf?: string
    cep?: string
    logradouro?: string
    numero?: string
    bairro?: string
    localidade?: string
    selecionado?: boolean
    uf?: string,
    s1?: number,
    s2?: number,
    s3?: number,
    TotalRepasse?: number,
    valorTotal?: number
}

 